engines: rename pkg to engine
