+++ engine/engine_test.go
@@ -1,4 +1,4 @@
-package engines
+package engine
 
 import (
 	"testing"
